# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive property management API for real estate data",
        "title": "HomeInsight Properties API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
        "/login": {
            "post": {
                "description": "Authenticate user and return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/properties": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of all properties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Get all properties with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedPropertiesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new property record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Create a new property",
                "parameters": [
                    {
                        "description": "Property data",
                        "name": "property",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Property"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Property"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/properties/property-search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search for properties based on query string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Search for a specific property",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Property"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/properties/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single property by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Get property by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Property"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing property record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Update a property",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated property data",
                        "name": "property",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Property"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Property"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a property by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Delete a property",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Register a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                }
            }
        },
        "handlers.TokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "models.Address": {
            "type": "object",
            "properties": {
                "carrierRoute": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "county": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "streetAddress": {
                    "type": "string"
                },
                "streetAddressParsed": {
                    "$ref": "#/definitions/models.StreetAddressParsed"
                },
                "zipCode": {
                    "type": "string"
                },
                "zipPlus4": {
                    "type": "string"
                }
            }
        },
...
        "models.Building": {
            "type": "object",
            "properties": {
                "details": {
                    "$ref": "#/definitions/models.BuildingDetails"
                },
                "summary": {
                    "$ref": "#/definitions/models.BuildingSummary"
                }
            }
        },
        "models.BuildingDetails": {
            "type": "object",
            "properties": {
                "classification": {
                    "$ref": "#/definitions/models.Classification"
                },
                "construction": {
                    "$ref": "#/definitions/models.Construction"
                },
                "exterior": {
                    "$ref": "#/definitions/models.Exterior"
                },
                "interior": {
                    "$ref": "#/definitions/models.Interior"
                },
                "structureId": {
                    "$ref": "#/definitions/models.StructureID"
                },
                "verticalProfile": {
                    "$ref": "#/definitions/models.VerticalProfile"
                }
            }
        },
        "models.BuildingSummary": {
            "type": "object",
            "properties": {
                "bathroomFixturesCount": {
                    "type": "integer"
                },
                "bathroomsCount": {
                    "type": "integer"
                },
                "buildingsCount": {
                    "type": "integer"
                },
                "fireplacesCount": {
                    "type": "integer"
                },
                "fullBathroomsCount": {
                    "type": "integer"
                },
                "halfBathroomsCount": {
                    "type": "integer"
                },
                "livingAreaSquareFeet": {
                    "type": "integer"
                },
                "totalAreaSquareFeet": {
                    "type": "integer"
                }
            }
        },
        "models.Buyer": {
            "type": "object",
            "properties": {
                "firstNameAndMiddleInitial": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                }
            }
        },
        "models.CBSA": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.LandUseAndZoning": {
            "type": "object",
            "properties": {
                "landUseCode": {
                    "type": "string"
                },
                "propertyTypeCode": {
                    "type": "string"
                },
                "stateLandUseCode": {
                    "type": "string"
                },
                "stateLandUseDescription": {
                    "type": "string"
                }
            }
        },
        "models.LastMarketSale": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "bookNumber": {
                    "type": "string"
                },
                "buyers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Buyer"
                    }
                },
                "date": {
                    "type": "string"
                },
                "documentNumber": {
                    "type": "string"
                },
                "documentTypeCode": {
                    "type": "string"
                },
                "isMortgagePurchase": {
                    "type": "boolean"
                },
                "isResale": {
                    "type": "boolean"
                },
                "multiOrSplitParcelCode": {
                    "type": "string"
                },
                "pageNumber": {
                    "type": "string"
                },
                "recordingDate": {
                    "type": "string"
                },
                "sellers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Seller"
                    }
                },
                "titleCompany": {
                    "$ref": "#/definitions/models.TitleCompany"
                }
            }
        },
...
        "models.Location": {
            "type": "object",
            "properties": {
                "cbsa": {
                    "$ref": "#/definitions/models.CBSA"
                },
                "censusTract": {
                    "$ref": "#/definitions/models.CensusTract"
                },
                "coordinates": {
                    "$ref": "#/definitions/models.Coordinates"
                },
                "legal": {
                    "$ref": "#/definitions/models.Legal"
                }
            }
        },
        "models.Lot": {
            "type": "object",
            "properties": {
                "areaAcres": {
                    "type": "number"
                },
                "areaSquareFeet": {
                    "type": "integer"
                },
                "areaSquareFeetUsable": {
                    "type": "integer"
                },
                "topographyType": {
                    "type": "string"
                }
            }
        },
...
        "models.Ownership": {
            "type": "object",
            "properties": {
                "currentOwners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Owner"
                    }
                },
                "mailingAddress": {
                    "$ref": "#/definitions/models.MailingAddress"
                },
                "occupancyCode": {
                    "type": "string"
                },
                "relationshipTypeCode": {
                    "type": "string"
                }
            }
        },
        "models.PaginatedPropertiesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PropertyResponse"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/models.PaginationMeta"
                }
            }
        },
        "models.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "offset": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.Property": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "$ref": "#/definitions/models.Address"
                },
                "avmPropertyId": {
                    "type": "string"
                },
                "building": {
                    "$ref": "#/definitions/models.Building"
                },
                "landUseAndZoning": {
                    "$ref": "#/definitions/models.LandUseAndZoning"
                },
                "lastMarketSale": {
                    "$ref": "#/definitions/models.LastMarketSale"
                },
                "location": {
                    "$ref": "#/definitions/models.Location"
                },
                "lot": {
                    "$ref": "#/definitions/models.Lot"
                },
                "ownership": {
                    "$ref": "#/definitions/models.Ownership"
                },
                "propertyId": {
                    "type": "string"
                },
                "taxAssessment": {
                    "$ref": "#/definitions/models.TaxAssessment"
                },
                "utilities": {
                    "$ref": "#/definitions/models.Utilities"
                }
            }
        },
        "models.PropertyResponse": {
            "type": "object",
            "properties": {
                "property": {
                    "$ref": "#/definitions/models.Property"
                }
            }
        },
...
        "models.Seller": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                }
            }
        },
        "models.StreetAddressParsed": {
            "type": "object",
            "properties": {
                "houseNumber": {
                    "type": "string"
                },
                "streetName": {
                    "type": "string"
                },
                "streetNameSuffix": {
                    "type": "string"
                }
            }
        },
...
        "models.TaxAssessment": {
            "type": "object",
            "properties": {
                "assessedValue": {
                    "$ref": "#/definitions/models.AssessedValue"
                },
                "countyTaxAmount": {
                    "type": "integer"
                },
                "schoolDistrict": {
                    "$ref": "#/definitions/models.SchoolDistrict"
                },
                "taxRoll": {
                    "$ref": "#/definitions/models.TaxRoll"
                },
                "totalTaxAmount": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "models.Utilities": {
            "type": "object",
            "properties": {
                "electricityWiringTypeCode": {
                    "type": "string"
                },
                "fuelTypeCode": {
                    "type": "string"
                },
                "sewerTypeCode": {
                    "type": "string"
                },
                "utilitiesTypeCode": {
                    "type": "string"
                },
                "waterTypeCode": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}